requirements.txt
# Tweet Hours - Dependencies
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-dotenv==1.0.0
jinja2==3.1.3
httpx==0.26.0
playwright==1.41.0
beautifulsoup4==4.12.3
pydantic==2.5.3
